# openapi.yaml
openapi: 3.1.0
info:
  title: GravelExplorer API
  version: "1.0.0"
  description: >
    HTTP-API des GravelExplorer-Backends (Milestones 1–4) für die iOS-App.
    Bearer-Token-Auth. Alle Zeitstempel sind ISO-8601 UTC. IDs sind UUID-Strings.
    Erklärender Guide: siehe docs/API.md.
servers:
  - url: https://gravelexplorer.benx.de/api/v1
    description: Produktion
  - url: http://gravelexplorer.test/api/v1
    description: Lokal (MAMP)
...
security:
  - bearerAuth: []
paths:
...
  /discover/routes:
    get:
      tags: [Discovery]
      summary: Öffentliche Routen entdecken
      security:
        - bearerAuth: []
        - {}
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { $ref: '#/components/parameters/Sort' }
        - name: q
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: array, items: { type: string } }
        - name: bbox
          in: query
          description: "minLat,minLon,maxLat,maxLon"
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes: { type: array, items: { $ref: '#/components/schemas/Route' } }
                  pagination: { $ref: '#/components/schemas/Pagination' }
  /discover/users:
    get:
      tags: [Discovery]
      summary: Öffentliche Profile entdecken
      security:
        - bearerAuth: []
        - {}
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { type: array, items: { $ref: '#/components/schemas/PublicProfile' } }
                  pagination: { $ref: '#/components/schemas/Pagination' }
...
  /feed:
    get:
      tags: [Feed]
      summary: Routen gefolgter User (neueste zuerst)
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes: { type: array, items: { $ref: '#/components/schemas/Route' } }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        "401": { $ref: '#/components/responses/Error' }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Benachrichtigungen
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications: { type: array, items: { $ref: '#/components/schemas/Notification' } }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        "401": { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    Sort:
      name: sort
      in: query
      schema: { type: string, enum: [newest, oldest, distance_asc, distance_desc], default: newest }
...
  responses:
    Error:
      description: Fehler
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Visibility:
      type: string
      enum: [private, unlisted, public]
    Source:
      type: string
      enum: [app, import, strava, manual]
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
    Pagination:
      type: object
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
        has_more: { type: boolean }
...
    PublicProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string, nullable: true }
        public_handle: { type: string }
        created_at: { type: string, format: date-time }
        follower_count: { type: integer }
        following_count: { type: integer }
        route_count: { type: integer }
        is_self: { type: boolean }
        is_followed_by_viewer: { type: boolean }
      description: Felder können je nach Viewer variieren.
...
    LatLon:
      type: object
      properties:
        lat: { type: number }
        lon: { type: number }
    Route:
      type: object
      properties:
        id: { type: string, format: uuid }
        client_route_uuid: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string, nullable: true }
        visibility: { $ref: '#/components/schemas/Visibility' }
        source: { $ref: '#/components/schemas/Source' }
        version: { type: integer, nullable: true }
        format: { type: string, nullable: true, enum: [gpx, geojson] }
        stats:
          type: object
          properties:
            distance_m: { type: integer, nullable: true }
            elevation_gain_m: { type: integer, nullable: true }
            point_count: { type: integer, nullable: true }
            started_at: { type: string, format: date-time, nullable: true }
            ended_at: { type: string, format: date-time, nullable: true }
            bbox:
              nullable: true
              type: object
              properties:
                min_lat: { type: number }
                min_lon: { type: number }
                max_lat: { type: number }
                max_lon: { type: number }
            centroid:
              nullable: true
              allOf: [ { $ref: '#/components/schemas/LatLon' } ]
        tags:
          type: array
          items: { type: string }
        hints:
          type: array
          description: >-
            Wegpunkt-Hinweise zum Weg, die der Fahrer während der Aufzeichnung
            gesetzt hat (aus den GPX-<wpt>-Extensions geparst). Nur in der
            Einzelroute (GET /routes/{id}) enthalten, nicht in Listen.
          items: { $ref: '#/components/schemas/RouteHint' }
        owner:
          nullable: true
          type: object
          properties:
            handle: { type: string }
            display_name: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    RouteHint:
      type: object
      properties:
        reason_key:
          type: string
          description: "Grund-Schlüssel, ^[a-z0-9_]{1,40}$ (z. B. unrideable, great_view, custom_xxxx)"
        sentiment: { type: string, enum: [negative, positive] }
        label: { type: string, description: "Lesbarer Anzeigename zum Zeitpunkt der Erfassung" }
        note: { type: string, nullable: true }
        lat: { type: number }
        lon: { type: number }
        recorded_at: { type: string, format: date-time, nullable: true }
...
    Notification:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [follow, like, comment] }
        created_at: { type: string, format: date-time }
        read: { type: boolean }
        actor:
          type: object
          properties:
            handle: { type: string }
            display_name: { type: string, nullable: true }
        route:
          nullable: true
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
...